refactor: split game logic utils
