+++ tests/inventoryUtils.test.ts
@@ -1,0 +1,55 @@
+import { describe, it, expect } from 'vitest';
+import { applyItemChangeAction, buildItemChangeRecords, applyAllItemChanges } from '../utils/inventoryUtils';
+import { PLAYER_HOLDER_ID } from '../constants';
+import type { ItemChange } from '../types';
+
+describe('inventoryUtils', () => {
+  it('applyItemChangeAction adds gained item', () => {
+    const change: ItemChange = {
+      action: 'gain',
+      item: { id: 'it1', name: 'Torch', type: 'equipment', description: 'Bright', holderId: PLAYER_HOLDER_ID },
+    };
+    const result = applyItemChangeAction([], change);
+    expect(result).toHaveLength(1);
+    expect(result[0].name).toBe('Torch');
+  });
+
+  it('buildItemChangeRecords returns gain record', () => {
+    const change: ItemChange = {
+      action: 'gain',
+      item: { id: 'it1', name: 'Torch', type: 'equipment', description: 'Bright', holderId: PLAYER_HOLDER_ID },
+    };
+    const records = buildItemChangeRecords([change], []);
+    expect(records).toEqual([
+      {
+        type: 'gain',
+        gainedItem: {
+          id: 'it1',
+          name: 'Torch',
+          type: 'equipment',
+          description: 'Bright',
+          activeDescription: undefined,
+          isActive: false,
+          isJunk: false,
+          knownUses: [],
+          holderId: PLAYER_HOLDER_ID,
+        },
+      },
+    ]);
+  });
+
+  it('applyAllItemChanges applies multiple changes', () => {
+    const changes: Array<ItemChange> = [
+      {
+        action: 'gain',
+        item: { id: 'it1', name: 'Torch', type: 'equipment', description: 'Bright', holderId: PLAYER_HOLDER_ID },
+      },
+      {
+        action: 'update',
+        item: { id: 'it1', name: 'Torch', isActive: true, holderId: PLAYER_HOLDER_ID },
+      },
+    ];
+    const result = applyAllItemChanges(changes, []);
+    expect(result[0].isActive).toBe(true);
+  });
+});
